requirements.txt
Flask==2.0.3
mysql-connector-python==8.0.27
Werkzeug==2.0.3
Flask-Login==0.6.0
gunicorn==20.1.0
Jinja2==3.0.3
itsdangerous==2.0.1
Flask-Mail==0.9.1
python-dotenv==1.1.0